Implement NTP Custom homepage
NTPBackgroundImagesService will control which background provider(SI or SR)
should be used for NTP background.
It's global service because SI or SR component is browser-wide component.
And ViewCounterService will handle profile-wide properties.

In this PR, kNewTabPageShowSuperReferrerBackgroundImage pref is added.
Client should use it as a switch for turn it on/off SR component in runtime.

Regarding to flags, existing New Tab Page Branded Wallpapers flag will act
as a top level switch. If it is off, SI and SR are all disabled.
If New Tab Page Branded Wallpapers feature switch is on, SI component is
enabled by default.

New "Enable Brave Custom Homepage" flag is introduced for SR component.
If this is on, SR component will be installed when current install comes
from super referrer. If not, SI component will be registered.

If SI flag and demo flag are on, super referrer demo data will be provided.
If SI flag is off and demo flags is on, sponsored images data will be provided.

Regarding to ntp-branded-data-path, this switch can be used for both.