[Android] TEMP: isSyncEnabled was removed Needs futher research to properly handle this. For now just commented to unblock build.
Chromium change:
https://chromium.googlesource.com/chromium/src/+/12166216f9de6248ae936722a3fac423349eb1e1

Make AndroidSyncSettings.isSyncEnabled() internal to SyncController

This method was one of the big villains of AndroidSyncSettings, because
it was easy to use it instead of some ProfileSyncService API. This CL
makes it internal to SyncController, avoiding new uses. A warning is
added to isChromeSyncEnabled() to prevent developers from thinking the
method has just been renamed. Extra care is taken not to weaken the
tests in AndroidSyncSettingsTest, replacing expectations on this method
accordingly.

Bug: 1107904