Use Chromium's upstream git repository instead of the GitHub mirror
The GitHub mirrow has been temporarily out-of-sync since past Wednesday
November 17th which, even though they run a couple of manual syncs in
the past days, is causing trouble for us to work on rebases because we
can't fetch the right tags from the mirror.

This change changes the URL in package.json to point to the upstream
repository and makes the necessary changes to npm run sync so that
the gclient and git configurations get updated to it.

According to the upstream bug at https://crbug.com/1272797, this
situation should be fixed soon as it's been treated as high prio,
but we can't depend on that to be able to update to newer tags, so
we'd better make this change now, and then maybe return to using the
GitHub mirror later on when the situation is fixed, if needed.