Remove remaining pointer fields in brave
This change corrects all sites where pointer fields were still remaining
in the codebase, and replaces them with `raw_ptr<T>`, and `raw_ref<T>`.
This is in accordance with upstream hardening requirements, which are
about to be enforced by chromium's clang plugin, in M114.

This change goes one step further, and replaces pointer with references
in types where nullptr values are not supposed to be supported.