[Android] Tab Model to use TabClosureParams
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/2bec252dc800a55defad9247c0e1f7f98c8df86c

[Tab Model] Use TabClosureParams

This CL deprecates all the flavors of tab closure in the TabModel in
favor of a single `closeTabs` method with a set of params.

There are no logical changes inside TabModelImpl in this CL although it
should now be possible to steamline the flow if we so choose.

The main impetus for this change is that we will be introducing warning
in certain contexts when closing a tab may result in loss of data.
Having a single consistent API for closing tabs will allow us to
introduce patterns such as the one below without needing to awkwardly
forward args and dispatch to the right method on TabModel.

Additionally the TabClosureParams are a lot cleaner, more extensible,
and easier to reason about than lists of several booleans in a row as
was the previous the case. For example; one might consider introducing
a close tab group param using the TabGroupId as a key rather than a
list of tabs. With this new API that is realtively easy and the
TabGroupId can be resolve to a list of tabs in the TabModelImpl.

```
closeTabsWithPossibleWarning(TabClosureParams params) {
    if (!shouldWarn(params)) {
        tabModel.closeTabs(params);
	return;
    } else if (shouldCreateNtp(params)) {
        // Prevent a tab group from closing by adding an NTP.
	createTab();
	tabModel.closeTabs(params);
	return;
    }
    Callback<Integer> dialogResult = (result) -> {
        if (result == ACCEPT) {
            tabModel.closeTabs(params);
        }
    };
    ModalDialogManager.showDialog(...);
}
```

Low-Coverage-Reason: LARGE_SCALE_REFACTOR only ChromeTabbedActivity is lacking coverage.
Fixed: 356445932