//chrome/common/extensions build target was added
Chromium change:

https://chromium.googlesource.com/chromium/src/+/ec463f3e1ee7b

commit ec463f3e1ee7b6f66b4c71d0dbd772c4c8d07ad2
Author: Devlin Cronin <rdevlin.cronin@chromium.org>
Date:   Fri Aug 16 02:24:51 2024 +0000

    [Extensions] Add a //chrome/common/extensions build target

    Extract out files from //chrome/common/extensions into a separate
    build target, //chrome/common/extensions:extensions, instead of having
    them shared in //chrome/common:common_lib.

    In addition to reducing the monolithic build file in //chrome/common,
    this some circular dependencies between //chrome/common resources and
    extensions files.

    Bug: 356905053

Fixes dependency on //brave/common/extensions/api errors:

ERROR at //brave/common/extensions/api/generated_includes.h:8:11: Can't include this header from here.
          ^-----------------------------------------
The target:
  //chrome/common/extensions:extensions
is including a file from the target:
  //brave/common/extensions/api:api_features_json_features

It's usually best to depend directly on the destination target.
In some cases, the destination target is considered a subcomponent
of an intermediate target. In this case, the intermediate target
should depend publicly on the destination to forward the ability
to include headers.

Dependency chain (there may also be others):
  //chrome/common/extensions:extensions -->
  //brave/common/extensions/api:api -->
  //brave/common/extensions/api:generated_api_json_strings --[private]-->
  //brave/common/extensions/api:api_features -->
  //brave/common/extensions/api:api_features_json_features

___________________
ERROR at //brave/common/extensions/api/generated_includes.h:9:11: Can't include this header from here.
          ^----------------------------------------------
The target:
  //chrome/common/extensions:extensions
is including a file from the target:
  //brave/common/extensions/api:behavior_features_json_features

It's usually best to depend directly on the destination target.
In some cases, the destination target is considered a subcomponent
of an intermediate target. In this case, the intermediate target
should depend publicly on the destination to forward the ability
to include headers.

Dependency chain (there may also be others):
  //chrome/common/extensions:extensions -->
  //brave/common/extensions/api:api -->
  //brave/common/extensions/api:generated_api_json_strings --[private]-->
  //brave/common/extensions/api:behavior_features -->
  //brave/common/extensions/api:behavior_features_json_features

___________________
ERROR at //brave/common/extensions/api/generated_includes.h:12:11: Can't include this header from here.
          ^-----------------------------------------------------
The target:
  //chrome/common/extensions:extensions
is including a file from the target:
  //brave/common/extensions/api:resources_grit

It's usually best to depend directly on the destination target.
In some cases, the destination target is considered a subcomponent
of an intermediate target. In this case, the intermediate target
should depend publicly on the destination to forward the ability
to include headers.

Dependency chain (there may also be others):
  //chrome/common/extensions:extensions -->
  //brave/common/extensions/api:api -->
  //brave/common/extensions/api:generated_api_json_strings --[private]-->
  //brave/common/extensions/api:resources -->
  //brave/common/extensions/api:resources_grit

___________________
ERROR at //brave/common/extensions/api/generated_includes.h:13:11: Can't include this header from here.
          ^----------------------------------------------
The target:
  //chrome/common/extensions:extensions
is including a file from the target:
  //brave/common/extensions/api:manifest_features_json_features

It's usually best to depend directly on the destination target.
In some cases, the destination target is considered a subcomponent
of an intermediate target. In this case, the intermediate target
should depend publicly on the destination to forward the ability
to include headers.

Dependency chain (there may also be others):
  //chrome/common/extensions:extensions -->
  //brave/common/extensions/api:api -->
  //brave/common/extensions/api:generated_api_json_strings --[private]-->
  //brave/common/extensions/api:manifest_features -->
  //brave/common/extensions/api:manifest_features_json_features

___________________
ERROR at //brave/common/extensions/api/generated_includes.h:14:11: Can't include this header from here.
          ^------------------------------------------------
The target:
  //chrome/common/extensions:extensions
is including a file from the target:
  //brave/common/extensions/api:permission_features_json_features

It's usually best to depend directly on the destination target.
In some cases, the destination target is considered a subcomponent
of an intermediate target. In this case, the intermediate target
should depend publicly on the destination to forward the ability
to include headers.

Dependency chain (there may also be others):
  //chrome/common/extensions:extensions -->
  //brave/common/extensions/api:api -->
  //brave/common/extensions/api:generated_api_json_strings --[private]-->
  //brave/common/extensions/api:permission_features -->
  //brave/common/extensions/api:permission_features_json_features