`SyncService::StopAndClear` deleted
This event used to be part of the base class. Now with its deletion the
internal function `SyncServiceImpl::StopAndClear` has to be exposed
through `BraveSyncServiceImpl` to keep the existing code in use.

This change turns the private function into a virtual function, to allow
us to keep plugging our own changes into it.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/769c49bea8b02913b69c1607bbe3d2190a3067bb

commit 769c49bea8b02913b69c1607bbe3d2190a3067bb
Author: Victor Hugo Vianna Silva <victorvianna@google.com>
Date:   Wed Aug 21 12:41:28 2024 +0000

    Remove SyncService::StopAndClear()

    The method can be removed from the interface, its last usage was
    removed in crrev.com/c/5800572.

    Bug: 40797392