[WIP] Fixes dependency in common/extensions/api.
generated_api_registration doesn't belong in common/extensions/api
because it should depend on the apis implementations that are done
in browser/extensions.

Moved it to browser/extensions (ideally we should move it to
browser/extensions/api and have a target there that builds
implementations, but leaving that for later).

Fixes a lot of errors similar to:

ld.lld: error: undefined symbol: vtable for extensions::api::BraveRewardsIsSupportedFunction
>>> referenced by brave_rewards_api.h:21 (../../brave/browser/extensions/api/brave_rewards_api.h:21)
>>>               generated_api_registration/generated_api_registration.o:(extensions::api::BraveRewardsIsSupportedFunction::BraveRewardsIsSupportedFunction()) in archive obj/brave/common/extensions/api/libgenerated_api_registration.a
>>> the vtable symbol may be undefined because the class is missing its key function (see https://lld.llvm.org/missingkeyfunction)